app/requirements.txt
semantic-kernel
azure-identity
azure-keyvault-secrets
azure-search-documents
azure-core